Actually support looking at favourites
Minimal support for actually looking at favourites through the UI.  Many things
are still wrong, but this is already a huge commit.

Added a new Activity for viewing the favourite, and a bunch of database stuff
for fetching the relevant info from the DB.

Had to update the database generation scripts because we needed a better way to
figure out the schedule for each day.