Get database from web rather than assets
Putting the full database in the assets increased the APK size by 10s of MB,
and means that databse updates entail an app update.  Better to download the
updates from the web.

For now I am hosting the file on github, where we can (and do) use etags to
detect whether a new version is available.